backend/requirements.txt
Django==3.1.7
psycopg2-binary==2.8.6
djangorestframework==3.12.2
djangorestframework-camel-case==1.2.0
